tạm hoàn thành admin

+++ app/src/main/java/com/seo/visualexhibition/data/model/Topic.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "topics")
 data class Topic(
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = "topic_id") val id: Int = 0,
+    @ColumnInfo(name = "topic_id") val id: Long = 0,
     @ColumnInfo(name = "topic_name") val topicName: String,
     @ColumnInfo(name = "topic_description") val topicDescription: String?,
     @ColumnInfo(name = "image_src") val imageSrc: String
