+++ app/src/main/java/com/seo/visualexhibition/data/model/User.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "users")
 data class User(
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = "user_id") val id: Int,
+    @ColumnInfo(name = "user_id") val id: Long,
     @ColumnInfo(name = "username") val username: String,
     @ColumnInfo(name = "password") val password: String,
     @ColumnInfo(name = "role") val role: String
